Project structure for separating unit test from integration test. Added gradle plugin for having beautiful test logs. Application moved to infrastructure package, we are going to use hexagonal architecture.
