+++ src/integrationTest/java/org/happyhome/myorhcard/infrastructure/MyorhcardApplicationTests.java
@@ -1,4 +1,4 @@
-package org.happyhome.myorhcard;
+package org.happyhome.myorhcard.infrastructure;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
