+++ src/test/java/org/happyhome/myorhcard/UnitTest.java
@@ -1,0 +1,15 @@
+package org.happyhome.myorhcard;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class UnitTest {
+
+  @Test
+  void unitTest() {
+    int a=1;
+    int b=1;
+    assertThat(a).isEqualTo(b);
+  }
+}
